+++ meringue/api/__init__.py
@@ -1,0 +1,5 @@
+import django
+
+
+if django.VERSION < (3, 2):
+    default_app_config = "meringue.api.apps.Config"