+++ meringue/api/docs/__init__.py
@@ -1,0 +1,2 @@
+from meringue.api.docs.patchers import *
+from meringue.api.docs.views import *