+++ meringue/core/query.py
@@ -4,6 +4,10 @@ from django.utils import timezone
 
 
 class SortingQuerySet(QuerySet):
+    """
+    QuerySet with methods for working with sorted positions
+    """
+
     def correction_sorting(self):
         """
         This is a method to update/fix the sorting of the selected list of items.
@@ -25,40 +29,50 @@ class SortingQuerySet(QuerySet):
 
 
 class PublicationQuerySet(QuerySet):
+    """
+    QuerySet with methods for fetching published and unpublished elements.
+    """
+
     def published(self, *args, **kwargs):
         """
         Method to getting published items.
         """
-        kwargs["is_published"] = True
+        from meringue.core.models import PublicationDatesMixin
+        from meringue.core.models import PublicationMixin
+
+        if issubclass(self.model, PublicationMixin):
+            kwargs["is_published"] = True
+
+        if issubclass(self.model, PublicationDatesMixin):
+            now = timezone.now()
+            args = (
+                *args,
+                Q(date_from__lte=now, date_to__gt=now)
+                | Q(date_from__isnull=True, date_to__gt=now)
+                | Q(date_from__lte=now, date_to__isnull=True)
+                | Q(date_from__isnull=True, date_to__isnull=True),
+            )
+
         return self.filter(*args, **kwargs)
 
     def unpublished(self, *args, **kwargs):
         """
         Method for getting unpublished items.
         """
-        kwargs["is_published"] = False
+        from meringue.core.models import PublicationDatesMixin
+        from meringue.core.models import PublicationMixin
+
+        if issubclass(self.model, PublicationMixin):
+            kwargs["is_published"] = False
+
+        if issubclass(self.model, PublicationDatesMixin):
+            now = timezone.now()
+            args = (*args, Q(date_from__gt=now) | Q(date_to__lte=now))
+
         return self.filter(*args, **kwargs)
 
 
-class PublicationDatesQuerySet(QuerySet):
-    def published(self, *args, **kwargs):
-        """
-        Method to getting published items.
-        """
-        now = timezone.localtime()
-        args = (
-            *args,
-            Q(date_from__lte=now, date_to__gt=now)
-            | Q(date_from__isnull=True, date_to__gt=now)
-            | Q(date_from__lte=now, date_to__isnull=True)
-            | Q(date_from__isnull=True, date_to__isnull=True),
-        )
-        return self.filter(*args, **kwargs)
-
-    def unpublished(self, *args, **kwargs):
-        """
-        Method for getting unpublished items.
-        """
-        now = timezone.localtime()
-        args = (*args, Q(date_from__gt=now) | Q(date_to__lte=now))
-        return self.filter(*args, **kwargs)
+class UniversalQuerySet(SortingQuerySet, PublicationQuerySet):
+    """
+    Generic QuerySet containing methods of all other QuerySets
+    """